requirements.txt
# Video processing
av==10.0.0
opencv-python==4.8.0.76
pymediainfo==6.0.0

# OCR API
requests==2.31.0

# Text processing
thefuzz==0.19.0
wordninja-enhanced==1.0.0
opencc-python-reimplemented==0.1.7
fast-ssim==0.1.0

# System
wakepy==0.8.0
cpuid==0.1.0

# Testing
pytest==7.4.3
